+++ assembly/assembly-wsmaster-war/pom.xml
@@ -152,10 +152,6 @@
         </dependency>
         <dependency>
             <groupId>com.codenvy.onpremises.wsmaster</groupId>
-            <artifactId>codenvy-hosted-workspace-activity</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.codenvy.onpremises.wsmaster</groupId>
             <artifactId>codenvy-ldap</artifactId>
         </dependency>
         <dependency>
@@ -165,6 +161,10 @@
         <dependency>
             <groupId>com.codenvy.plugin</groupId>
             <artifactId>codenvy-machine-hosted</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.codenvy.plugin</groupId>
+            <artifactId>codenvy-plugin-activity-wsmaster</artifactId>
         </dependency>
         <dependency>
             <groupId>com.codenvy.plugin</groupId>
