+++ assembly/compiling-ide-war/pom.xml
@@ -33,6 +33,10 @@
         </dependency>
         <dependency>
             <groupId>com.codenvy.plugin</groupId>
+            <artifactId>codenvy-plugin-activity-ide</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.codenvy.plugin</groupId>
             <artifactId>codenvy-plugin-bitbucket-ext-bitbucket-client</artifactId>
         </dependency>
         <dependency>
@@ -58,6 +62,10 @@
             <exclusions>
                 <exclusion>
                     <artifactId>che-plugin-product-info</artifactId>
+                    <groupId>org.eclipse.che.plugin</groupId>
+                </exclusion>
+                <exclusion>
+                    <artifactId>che-plugin-activity-ide</artifactId>
                     <groupId>org.eclipse.che.plugin</groupId>
                 </exclusion>
             </exclusions>
