+++ plugins/plugin-activity/codenvy-plugin-activity-ide/src/main/resources/com/codenvy/plugin/activity/ActivityModule.gwt.xml
@@ -1,0 +1,19 @@
+<!--
+
+    Copyright (c) [2012] - [2017] Red Hat, Inc.
+    All rights reserved. This program and the accompanying materials
+    are made available under the terms of the Eclipse Public License v1.0
+    which accompanies this distribution, and is available at
+    http://www.eclipse.org/legal/epl-v10.html
+
+    Contributors:
+      Red Hat, Inc. - initial API and implementation
+
+-->
+<module>
+    <inherits name="com.google.gwt.user.User"/>
+
+    <public path="public"/>
+
+    <source path="ide"/>
+</module>
