--- plugins/plugin-hosted/codenvy-hosted-ext-hosted/src/main/java/com/codenvy/ide/hosted/client/ActivityTrackingExtension.java
@@ -1,57 +1,0 @@
-/*******************************************************************************
- * Copyright (c) [2012] - [2017] Red Hat, Inc.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *
- * Contributors:
- *   Red Hat, Inc. - initial API and implementation
- *******************************************************************************/
-package com.codenvy.ide.hosted.client;
-
-import com.google.gwt.core.client.Callback;
-import com.google.gwt.core.client.ScriptInjector;
-import com.google.inject.Singleton;
-
-import org.eclipse.che.ide.api.app.AppContext;
-import org.eclipse.che.ide.api.extension.Extension;
-import org.eclipse.che.ide.rest.RestContext;
-
-import javax.inject.Inject;
-
-import static com.google.gwt.core.client.ScriptInjector.TOP_WINDOW;
-
-/**
- * Adds activity tracking script to IDE.
- *
- * @author Max Shaposhnik (mshaposhnik@codenvy.com)
- */
-@Singleton
-@Extension(title = "Activity Tracking Extension", version = "1.0.0")
-public class ActivityTrackingExtension {
-
-    @Inject
-    public ActivityTrackingExtension(final @RestContext String restContext,
-                                     final AppContext appContext) {
-
-        ScriptInjector.fromUrl("/_app/activity.js")
-                      .setWindow(TOP_WINDOW)
-                      .setCallback(new Callback<Void, Exception>() {
-                          @Override
-                          public void onSuccess(Void result) {
-                              init(restContext, appContext.getWorkspaceId());
-                          }
-
-                          @Override
-                          public void onFailure(Exception reason) {
-                          }
-
-                          private native void init(String restContext, String wsId) /*-{
-                              $wnd.ActivityTracker.init(restContext, wsId);
-                          }-*/;
-                      })
-                      .inject();
-
-    }
-}
