+++ plugins/plugin-hosted/codenvy-hosted-ext-hosted/src/main/java/com/codenvy/ide/hosted/client/inject/HostedEnvironmentGinModule.java
@@ -15,8 +15,6 @@ import com.codenvy.ide.hosted.client.login.PromptToLoginView;
 import com.codenvy.ide.hosted.client.login.PromptToLoginViewImpl;
 import com.codenvy.ide.hosted.client.notifier.BadConnectionNotifierView;
 import com.codenvy.ide.hosted.client.notifier.BadConnectionNotifierViewImpl;
-import com.codenvy.ide.hosted.client.workspace.WorkspaceNotRunningView;
-import com.codenvy.ide.hosted.client.workspace.WorkspaceNotRunningViewImpl;
 import com.google.gwt.inject.client.AbstractGinModule;
 
 import org.eclipse.che.ide.api.extension.ExtensionGinModule;
@@ -32,7 +30,5 @@ public class HostedEnvironmentGinModule extends AbstractGinModule {
         bind(ConnectionClosedInformerImpl.class).to(HostedEnvConnectionClosedInformer.class).in(javax.inject.Singleton.class);
         bind(PromptToLoginView.class).to(PromptToLoginViewImpl.class);
         bind(BadConnectionNotifierView.class).to(BadConnectionNotifierViewImpl.class);
-        bind(WorkspaceNotRunningView.class).to(WorkspaceNotRunningViewImpl.class);
-//        bind(HostedWorkspaceStoppedHandler.class).asEagerSingleton();
     }
 }
