--- plugins/plugin-hosted/codenvy-hosted-ext-hosted/src/main/java/com/codenvy/ide/hosted/client/workspace/WorkspaceNotRunningView.java
@@ -1,30 +1,0 @@
-/*******************************************************************************
- * Copyright (c) [2012] - [2017] Red Hat, Inc.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *
- * Contributors:
- *   Red Hat, Inc. - initial API and implementation
- *******************************************************************************/
-package com.codenvy.ide.hosted.client.workspace;
-
-/**
- * @author Mihail Kuznyetsov
- */
-public interface WorkspaceNotRunningView {
-
-
-    interface ActionDelegate {
-        /** Performs some actions when user clicks on open dashboard button. */
-        void onOpenDashboardButtonClicked();
-
-        /** Performs some actions when user clicks on restart button. */
-        void onRestartButtonClicked();
-    }
-
-    void show();
-
-    void setDelegate(ActionDelegate delegate);
-}
