--- plugins/plugin-hosted/codenvy-hosted-ext-hosted/src/main/java/com/codenvy/ide/hosted/client/workspace/WorkspaceNotRunningViewImpl.ui.xml
@@ -1,25 +1,0 @@
-<!--
-
-    Copyright (c) [2012] - [2017] Red Hat, Inc.
-    All rights reserved. This program and the accompanying materials
-    are made available under the terms of the Eclipse Public License v1.0
-    which accompanies this distribution, and is available at
-    http://www.eclipse.org/legal/epl-v10.html
-
-    Contributors:
-      Red Hat, Inc. - initial API and implementation
-
--->
-<ui:UiBinder xmlns:ui='urn:ui:com.google.gwt.uibinder'
-             xmlns:g='urn:import:com.google.gwt.user.client.ui'>
-
-    <ui:with field="locale" type="com.codenvy.ide.hosted.client.HostedLocalizationConstant"/>
-
-    <ui:style>
-        .label {
-            margin: 10px;
-        }
-    </ui:style>
-
-    <g:Label width="450px" addStyleNames="{style.label}" text="{locale.workspaceNotRunningMessage}" ui:field="label"/>
-</ui:UiBinder>
