--- wsmaster/codenvy-hosted-workspace-activity-shared/pom.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright (c) [2012] - [2017] Red Hat, Inc.
-    All rights reserved. This program and the accompanying materials
-    are made available under the terms of the Eclipse Public License v1.0
-    which accompanies this distribution, and is available at
-    http://www.eclipse.org/legal/epl-v10.html
-
-    Contributors:
-      Red Hat, Inc. - initial API and implementation
-
--->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <artifactId>codenvy-master-parent</artifactId>
-        <groupId>com.codenvy.onpremises.wsmaster</groupId>
-        <version>5.16.0-SNAPSHOT</version>
-    </parent>
-    <artifactId>codenvy-hosted-workspace-activity-shared</artifactId>
-    <packaging>jar</packaging>
-    <name>Codenvy :: Hosted :: Workspace Activity Shared</name>
-</project>
