--- wsmaster/codenvy-hosted-workspace-activity-shared/src/main/java/com/codenvy/activity/shared/Constants.java
@@ -1,24 +1,0 @@
-/*******************************************************************************
- * Copyright (c) [2012] - [2017] Red Hat, Inc.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *
- * Contributors:
- *   Red Hat, Inc. - initial API and implementation
- *******************************************************************************/
-package com.codenvy.activity.shared;
-
-/**
- * Workspace activity constants.
- *
- * @author Mihail Kuznyetsov
- */
-public final class Constants {
-
-    public static final String ACTIVITY_CHECKER     = "activity-checker";
-
-    private Constants() {
-    }
-}
