--- wsmaster/codenvy-hosted-workspace-activity/src/main/java/com/codenvy/activity/server/inject/WorkspaceActivityModule.java
@@ -1,28 +1,0 @@
-/*******************************************************************************
- * Copyright (c) [2012] - [2017] Red Hat, Inc.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *
- * Contributors:
- *   Red Hat, Inc. - initial API and implementation
- *******************************************************************************/
-package com.codenvy.activity.server.inject;
-
-import com.codenvy.activity.server.ActivityPermissionsFilter;
-import com.codenvy.activity.server.HostedWorkspaceActivityManager;
-import com.google.inject.AbstractModule;
-
-import org.eclipse.che.api.workspace.server.activity.WorkspaceActivityManager;
-import org.eclipse.che.api.workspace.server.activity.WorkspaceActivityService;
-
-public class WorkspaceActivityModule extends AbstractModule {
-
-    @Override
-    protected void configure() {
-        bind(ActivityPermissionsFilter.class);
-        bind(WorkspaceActivityService.class);
-        bind(WorkspaceActivityManager.class).to(HostedWorkspaceActivityManager.class);
-    }
-}
