--- wsmaster/codenvy-hosted-workspace-activity/src/test/java/com/codenvy/activity/server/ActivityPermissionsFilterTest.java
@@ -1,111 +1,0 @@
-/*******************************************************************************
- * Copyright (c) [2012] - [2017] Red Hat, Inc.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *
- * Contributors:
- *   Red Hat, Inc. - initial API and implementation
- *******************************************************************************/
-package com.codenvy.activity.server;
-
-import com.jayway.restassured.response.Response;
-
-import org.eclipse.che.api.core.ForbiddenException;
-import org.eclipse.che.api.workspace.server.activity.WorkspaceActivityService;
-import org.eclipse.che.commons.env.EnvironmentContext;
-import org.eclipse.che.commons.subject.Subject;
-import org.everrest.assured.EverrestJetty;
-import org.everrest.core.Filter;
-import org.everrest.core.GenericContainerRequest;
-import org.everrest.core.RequestFilter;
-import org.everrest.core.resource.GenericResourceMethod;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.Mockito;
-import org.mockito.testng.MockitoTestNGListener;
-import org.testng.annotations.Listeners;
-import org.testng.annotations.Test;
-
-
-import static com.codenvy.api.workspace.server.WorkspaceDomain.DOMAIN_ID;
-import static com.codenvy.api.workspace.server.WorkspaceDomain.USE;
-import static com.jayway.restassured.RestAssured.given;
-import static org.everrest.assured.JettyHttpServer.ADMIN_USER_NAME;
-import static org.everrest.assured.JettyHttpServer.ADMIN_USER_PASSWORD;
-import static org.everrest.assured.JettyHttpServer.SECURE_PATH;
-import static org.mockito.Matchers.anyString;
-import static org.mockito.Matchers.eq;
-import static org.mockito.Mockito.doThrow;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-import static org.testng.Assert.assertEquals;
-
-/**
- * Tests for {@link ActivityPermissionsFilter}.
- *
- * @author Max Shaposhnik
- */
-@Listeners(value = {EverrestJetty.class, MockitoTestNGListener.class})
-public class ActivityPermissionsFilterTest {
-
-    @SuppressWarnings("unused")
-    private static final EnvironmentFilter FILTER = new EnvironmentFilter();
-
-    @SuppressWarnings("unused")
-    @InjectMocks
-    ActivityPermissionsFilter permissionsFilter;
-
-    @Mock
-    private static Subject subject;
-
-    @Mock
-    WorkspaceActivityService service;
-
-    @Test
-    public void shouldCheckPermissionsOnGettingMachineById() throws Exception {
-
-        when(subject.hasPermission(eq(DOMAIN_ID), eq("workspace123"), eq(USE))).thenReturn(true);
-
-        final Response response = given().auth()
-                                         .basic(ADMIN_USER_NAME, ADMIN_USER_PASSWORD)
-                                         .when()
-                                         .put(SECURE_PATH + "/activity/workspace123");
-
-        assertEquals(response.getStatusCode(), 204);
-        verify(service).active(eq("workspace123"));
-        verify(subject).checkPermission(DOMAIN_ID, "workspace123", USE);
-    }
-
-    @Test
-    public void shouldThrowExceptionWhenUpdatingNotOwnedWorkspace() throws Exception {
-
-        when(subject.hasPermission(eq(DOMAIN_ID), eq("workspace123"), eq(USE))).thenReturn(false);
-        doThrow(new ForbiddenException("The user does not have permission to " + USE + " workspace with id 'workspace123'"))
-                .when(subject).checkPermission(anyString(), anyString(), anyString());
-
-        final Response response = given().auth()
-                                         .basic(ADMIN_USER_NAME, ADMIN_USER_PASSWORD)
-                                         .when()
-                                         .put(SECURE_PATH + "/activity/workspace123");
-
-        assertEquals(response.getStatusCode(), 403);
-    }
-
-    @Test(expectedExceptions = ForbiddenException.class)
-    public void shouldThrowExceptionWhenCallingUnlistedMethod() throws Exception {
-
-        GenericResourceMethod genericResourceMethod = Mockito.mock(GenericResourceMethod.class);
-        when(genericResourceMethod.getMethod()).thenReturn(this.getClass().getDeclaredMethod("shouldThrowExceptionWhenCallingUnlistedMethod"));
-        Object[] argument = new Object[0];
-        permissionsFilter.filter(genericResourceMethod, argument);
-    }
-
-    @Filter
-    public static class EnvironmentFilter implements RequestFilter {
-        public void doFilter(GenericContainerRequest request) {
-            EnvironmentContext.getCurrent().setSubject(subject);
-        }
-    }
-}
